From https://groups.google.com/forum/#!topic/comp.sys.apple2/NWjJ1_ytewQ (Thanks fatdog)
Kent Dickey writes:
Fantavision hangs when it starts on KEGS-based emulators (which includes
GSport and GSplus) due to a bug in the way KEGS handles the WAI instruction.
Fantavision masks interrupts with SEI, sets a SHR line interrupt, and then
does a WAI instruction. This hangs KEGS since it expects an actual interrupt
to end the WAI, but there won't be one since interrupts are masked. I never
realized this was a valid WAI use case.